third_party/blink/web_tests/svg/as-image/svgview-references-use-counters-expected.txt

PASS internals.isUseCounted(document.querySelector('object').contentDocument, SVGSVGElementFragmentSVGView) is false
PASS internals.isUseCounted(document.querySelector('object').contentDocument, SVGSVGElementFragmentSVGViewElement) is false
PASS internals.isUseCounted(document.getElementById('svg-view').contentDocument, SVGSVGElementFragmentSVGView) is true
PASS internals.isUseCounted(document.getElementById('svg-view').contentDocument, SVGSVGElementFragmentSVGViewElement) is false
PASS internals.isUseCounted(document.getElementById('svg-view-element').contentDocument, SVGSVGElementFragmentSVGView) is false
PASS internals.isUseCounted(document.getElementById('svg-view-element').contentDocument, SVGSVGElementFragmentSVGViewElement) is true
PASS successfullyParsed is true

TEST COMPLETE

 
